+++ test/test_instruction_parsing.py
@@ -5,6 +5,8 @@ from bespokeasm.assembler.bytecode.word import Word
 from bespokeasm.assembler.label_scope import GlobalLabelScope
 from bespokeasm.assembler.label_scope import LabelScope
 from bespokeasm.assembler.label_scope import LabelScopeType
+from bespokeasm.assembler.label_scope.named_scope_manager import ActiveNamedScopeList
+from bespokeasm.assembler.label_scope.named_scope_manager import NamedScopeManager
 from bespokeasm.assembler.line_identifier import LineIdentifier
 from bespokeasm.assembler.line_object import LineObject
 from bespokeasm.assembler.line_object import LineWithWords
@@ -32,6 +34,7 @@ class TestInstructionParsing(unittest.TestCase):
         local_scope.set_label_value('.local_var', 10, 3)
         local_scope.set_label_value('.loop', 0x8020, 3)
         cls.label_values = local_scope
+        cls.active_named_scopes = ActiveNamedScopeList(NamedScopeManager())
 
     def setUp(self):
         InstructionLine._INSTRUCTUION_EXTRACTION_PATTERN = None
@@ -436,6 +439,7 @@ class TestInstructionParsing(unittest.TestCase):
             isa_model.default_origin,
             isa_model.predefined_memory_zones,
         )
+        active_named_scopes = ActiveNamedScopeList(NamedScopeManager())
 
         lineid = LineIdentifier(42, 'test_label_parsing')
         preprocessor = Preprocessor()
@@ -445,6 +449,7 @@ class TestInstructionParsing(unittest.TestCase):
             'LABEL = %00001111',
             isa_model,
             TestInstructionParsing.label_values,
+            active_named_scopes,
             memzone_mngr.global_zone,
             memzone_mngr,
             preprocessor,
@@ -460,6 +465,7 @@ class TestInstructionParsing(unittest.TestCase):
             '.local_label:',
             isa_model,
             TestInstructionParsing.label_values,
+            active_named_scopes,
             memzone_mngr.global_zone,
             memzone_mngr,
             preprocessor,
@@ -477,6 +483,7 @@ class TestInstructionParsing(unittest.TestCase):
             'old_style EQU 42H',
             isa_model,
             TestInstructionParsing.label_values,
+            active_named_scopes,
             memzone_mngr.global_zone,
             memzone_mngr,
             preprocessor,
@@ -631,6 +638,7 @@ class TestInstructionParsing(unittest.TestCase):
             isa_model.default_origin,
             isa_model.predefined_memory_zones,
         )
+        active_named_scopes = ActiveNamedScopeList(NamedScopeManager())
 
         lineid = LineIdentifier(33, 'test_numeric_bytecode_operand')
 
@@ -640,6 +648,7 @@ class TestInstructionParsing(unittest.TestCase):
         )
         t1.set_start_address(1)
         t1.label_scope = TestInstructionParsing.label_values
+        t1.active_named_scopes = active_named_scopes
         self.assertIsInstance(t1, InstructionLine)
         self.assertEqual(t1.word_count, 1, 'has 1 words')
         t1.generate_words()
@@ -655,6 +664,7 @@ class TestInstructionParsing(unittest.TestCase):
         )
         t2.set_start_address(1)
         t2.label_scope = TestInstructionParsing.label_values
+        t2.active_named_scopes = active_named_scopes
         self.assertIsInstance(t2, InstructionLine)
         self.assertEqual(t2.word_count, 1, 'has 1 words')
         t2.generate_words()
@@ -670,6 +680,7 @@ class TestInstructionParsing(unittest.TestCase):
         )
         t3.set_start_address(1)
         t3.label_scope = TestInstructionParsing.label_values
+        t3.active_named_scopes = active_named_scopes
         self.assertIsInstance(t3, InstructionLine)
         self.assertEqual(t3.word_count, 2, 'has 2 words')
         t3.generate_words()
@@ -685,6 +696,7 @@ class TestInstructionParsing(unittest.TestCase):
                 isa_model, memzone_mngr.global_zone, memzone_mngr,
             )
             e1.label_scope = TestInstructionParsing.label_values
+            e1.active_named_scopes = active_named_scopes
             e1.generate_words()
 
         with self.assertRaises(SystemExit, msg='test invalid enumerations'):
@@ -693,6 +705,7 @@ class TestInstructionParsing(unittest.TestCase):
                 isa_model, memzone_mngr.global_zone, memzone_mngr,
             )
             e2.label_scope = TestInstructionParsing.label_values
+            e2.active_named_scopes = active_named_scopes
             e2.generate_words()
 
     def test_numeric_enumeration_operand(self):
@@ -703,7 +716,7 @@ class TestInstructionParsing(unittest.TestCase):
             isa_model.default_origin,
             isa_model.predefined_memory_zones,
         )
-
+        active_named_scopes = ActiveNamedScopeList(NamedScopeManager())
         lineid = LineIdentifier(33, 'test_numeric_enumeration_operand')
 
         t1 = InstructionLine.factory(
@@ -712,6 +725,7 @@ class TestInstructionParsing(unittest.TestCase):
         )
         t1.set_start_address(1)
         t1.label_scope = TestInstructionParsing.label_values
+        t1.active_named_scopes = active_named_scopes
         self.assertIsInstance(t1, InstructionLine)
         self.assertEqual(t1.word_count, 1, 'has 1 words')
         t1.generate_words()
@@ -727,6 +741,7 @@ class TestInstructionParsing(unittest.TestCase):
         )
         t2.set_start_address(1)
         t2.label_scope = TestInstructionParsing.label_values
+        t2.active_named_scopes = active_named_scopes
         self.assertIsInstance(t2, InstructionLine)
         self.assertEqual(t2.word_count, 1, 'has 1 words')
         t2.generate_words()
@@ -742,6 +757,7 @@ class TestInstructionParsing(unittest.TestCase):
                 isa_model, memzone_mngr.global_zone, memzone_mngr,
             )
             e1.label_scope = TestInstructionParsing.label_values
+            e1.active_named_scopes = active_named_scopes
             e1.generate_words()
 
     def test_expressions_in_operations(self):
@@ -1067,7 +1083,7 @@ class TestInstructionParsing(unittest.TestCase):
             isa_model.word_size,
             isa_model.word_segment_size,
         )
-        value1 = bc1.get_value(TestInstructionParsing.label_values, 0x2010, 32)
+        value1 = bc1.get_value(TestInstructionParsing.label_values, TestInstructionParsing.active_named_scopes, 0x2010, 32)
         self.assertEqual(value1, 0x2020, 'byte code should match')
 
         # now, test argument value generation with LSB slicing
@@ -1084,11 +1100,11 @@ class TestInstructionParsing(unittest.TestCase):
             isa_model.word_size,
             isa_model.word_segment_size,
         )
-        value2 = bc2.get_value(TestInstructionParsing.label_values, 0x2010, 32)
+        value2 = bc2.get_value(TestInstructionParsing.label_values, TestInstructionParsing.active_named_scopes, 0x2010, 32)
         self.assertEqual(value2, 0x45, 'byte code should match')
 
         with self.assertRaises(ValueError, msg="address MSBs don't match"):
-            bc2.get_value(TestInstructionParsing.label_values, 0x2250, 32)
+            bc2.get_value(TestInstructionParsing.label_values, TestInstructionParsing.active_named_scopes, 0x2250, 32)
 
         # test 16-bit slize with a 32-bit address
         bc3 = AddressByteCodePart(
@@ -1104,10 +1120,10 @@ class TestInstructionParsing(unittest.TestCase):
             isa_model.word_size,
             isa_model.word_segment_size,
         )
-        value3 = bc3.get_value(TestInstructionParsing.label_values, 0x19451000, 32)
+        value3 = bc3.get_value(TestInstructionParsing.label_values, TestInstructionParsing.active_named_scopes, 0x19451000, 32)
         self.assertEqual(value3, 0x8899, 'byte code should match')
         with self.assertRaises(ValueError, msg="address MSBs don't match"):
-            bc3.get_value(TestInstructionParsing.label_values, 0x20241000, 32)
+            bc3.get_value(TestInstructionParsing.label_values, TestInstructionParsing.active_named_scopes, 0x20241000, 32)
 
         #  test error conditions
         # error case: extraneous comparison operators ignored
@@ -1133,6 +1149,7 @@ class TestInstructionParsing(unittest.TestCase):
         )
         t1.set_start_address(0x2F10)
         t1.label_scope = TestInstructionParsing.label_values
+        t1.active_named_scopes = TestInstructionParsing.active_named_scopes
         self.assertIsInstance(t1, InstructionLine)
         self.assertEqual(t1.word_count, 2, 'has 2 words')
         t1.generate_words()
@@ -1157,6 +1174,7 @@ class TestInstructionParsing(unittest.TestCase):
         )
         t2.set_start_address(0x2F10)
         t2.label_scope = TestInstructionParsing.label_values
+        t2.active_named_scopes = TestInstructionParsing.active_named_scopes
         self.assertIsInstance(t2, InstructionLine)
         self.assertEqual(t2.word_count, 3, 'has 3 words')
         t2.generate_words()
@@ -1177,6 +1195,7 @@ class TestInstructionParsing(unittest.TestCase):
         )
         t3.set_start_address(0x2F10)
         t3.label_scope = TestInstructionParsing.label_values
+        t3.active_named_scopes = TestInstructionParsing.active_named_scopes
         self.assertIsInstance(t3, InstructionLine)
         self.assertEqual(t3.word_count, 3, 'has 3 words')
         t3.generate_words()
@@ -1197,6 +1216,7 @@ class TestInstructionParsing(unittest.TestCase):
         )
         t4.set_start_address(0x2F10)
         t4.label_scope = TestInstructionParsing.label_values
+        t4.active_named_scopes = TestInstructionParsing.active_named_scopes
         self.assertIsInstance(t4, InstructionLine)
         self.assertEqual(t4.word_count, 3, 'has 3 words')
         with self.assertRaises(SystemExit, msg='address not in target memory zone'):